Update README.md
Should we use the more formal Panoramic vs. Pano if these are going to be public documents? I think using the more formal name might be better for helping people remember the name of the company and platform 

I also think we are a little inconsistent in our use of periods - some sentences end in periods and others do not.