Update platform-overivew.md (#4)
@maxcobert Is this correct "Dimensions - Technically defined as anything that is not a dimension, these can include entities..." Do you mean anything that is not a metric?

Dataset is sometimes capitalized and other times not capitalized - not sure which one you want but would recommend being consistent throughout

Overall need to be consistent if you are or are not putting periods at the ends of bullets. Right now it is a mixed bag - use one method throughout. Same with capitalization - either consistently capitalize the first word at each bullet point or do not