Update README.md (#10)
* Update README.md

This phrase "Pano was built on a simple observation. Businesses need to become more data-driven, but the deck is stacked against them." What does the deck stacked against them mean? I think this should be altered as it is a bit confusing. 

This phrase "* SQL is available, but not required". I thought SQL is required?

* limited

Co-authored-by: maxcobert <max.cobert@gmail.com>